+++ src/gauge/zerosum.jl
@@ -1,6 +1,10 @@
 zerosum(A::AbstractArray; dims = 1) = A .- mean(A; dims)
 zerosum!(A::AbstractArray; dims = 1) = A .= zerosum(A; dims)
 
+# zerosum only affects Potts layers
+has_potts_layers(rbm) =
+    rbm.visible isa Union{Potts,PottsGumbel} || rbm.hidden isa Union{Potts,PottsGumbel}
+
 """
     zerosum(rbm)